+++ src/feno/__main__.py
@@ -14,13 +14,20 @@ def main():
     parser.add_argument("--version", "-v", action="store_true", help="Prints the version")
 
     parser.add_argument("--brief", "-b", action="store_true", help="Brief mode")
-    parser.add_argument("--erase", "-e", action="store_true", help="Erase .html and .tio temp files")
-    parser.add_argument("--remote", "-r", action="store_true", help="Search for remote.cfg file and create absolute links")
+
     parser.add_argument("--tko", "-t", action="store_true", help="Insert tko preamble")
+    parser.add_argument("--remote", "-r", action="store_true", help="Search for remote.cfg file and create absolute links")
+    parser.add_argument("--erase", "-e", action="store_true", help="Erase .html and .tio temp files")
+
+    parser.add_argument("--full", "-f", action="store_true", help="Full mode - equivalent to -tre")
     
 
     args = parser.parse_args()
     Log.set_verbose(not args.brief)
+    if args.full:
+        args.tko = True
+        args.remote = True
+        args.erase = True
     
     if args.version:
         print(__version__)
@@ -35,7 +42,7 @@ def main():
 
         actions = Actions(target, args.remote, args.tko)
         Log.resume(hook, end=": [ ")
-        Log.verbose(hook, end=" ")
+        Log.verbose(hook)
 
         if not actions.validate():
             continue