+++ src/feno/actions.py
@@ -4,8 +4,8 @@ from .remote_md import RemoteMd, Title, RemoteCfg
 from .html import HTML
 from .cases import Cases
 from .log import Log
-from .tree import Tree
 from .mdpp import Mdpp
+from .filter import DeepFilter
 
 from typing import Optional
 import subprocess
@@ -58,7 +58,7 @@ class Actions:
     def need_rebuild(self):
         if Check.need_rebuild(self.source_dir, self.target):
             Log.resume("Changes ", end="")
-            Log.verbose(f"    Changes in {self.source_dir}")
+            Log.verbose(f"  Changes in {self.source_dir}")
             return True
             
         Log.verbose("")
@@ -74,32 +74,33 @@ class Actions:
                 print("\n    fail: proceeding without make absolute links")
             else:
                 found = True
-                Log.verbose(f"    remote.cfg: {cfg_path}")
+                Log.verbose(f"  remote.cfg: {cfg_path}")
                 cfg.read(cfg_path)
         RemoteMd.run(cfg, self.source_readme, self.remote_readme, self.hook, self.insert_tko_preamble)
         if self.make_remote and found:
             Log.resume("AbsoluteMd ", end="")
-        Log.verbose(f"    RemoteFile: {self.remote_readme}")
+        Log.verbose(f"  RemoteFile: {self.remote_readme}")
     
     # uses pandoc to generate html from markdown
     def html(self):
         title = Title.extract_title(self.source_readme)
         HTML.generate_html_with_pandoc(title, self.remote_readme, self.target_html, True)
         Log.resume("HTML ", end="")
-        Log.verbose(f"    HTML  file: {self.target_html}")
+        Log.verbose(f"  HTML  file: {self.target_html}")
 
     # uses tko to generate cases file
     def build_cases(self):
         Cases.run(self.cases, self.source_readme, self.source_dir)
         Log.resume("Cases ", end="")
-        Log.verbose(f"    Cases file: {self.cases}")
+        Log.verbose(f"  Cases file: {self.cases}")
 
     def copy_drafts(self):
         source_src = norm_join(self.source_dir, "src")
         if os.path.isdir(source_src):
             Log.resume("Drafts ", end="")
-            Log.verbose(f"    Drafts dir: {source_src}")
-            Tree.deep_filter_copy(source_src, self.cache_src, 5)
+            Log.verbose(f"  Drafts dir: {source_src}")
+            filter = DeepFilter().set_indent(4)
+            filter.copy(source_src, self.cache_src, 5)
 
     def run_local_sh(self):
         local_sh = norm_join(self.source_dir, "local.sh")
@@ -109,26 +110,26 @@ class Actions:
             subprocess.run("bash local.sh", shell=True)
             os.chdir(actual_chdir)
             Log.resume("Local.sh ", end="")
-            Log.verbose(f"    local.sh executed")
+            Log.verbose(f"  local.sh executed")
 
     def init_vpl(self):
         self.vpl = JsonVPL(self.title, open(self.target_html).read())
         self.vpl.set_cases(self.cases)
         if self.vpl.load_config_json(self.config_json, self.source_dir):
             Log.resume("Required ", end="")
-            Log.verbose(f"    CfgVplJson: {self.config_json}")
+            Log.verbose(f"  CfgVplJson: {self.config_json}")
         if self.vpl.load_drafts(self.cache_src):
             Log.resume("Drafts ", end="")
 
     def create_mapi(self):
         open(self.mapi_json, "w").write(str(self.vpl) + "\n")
         Log.resume("Mapi ", end="")
-        Log.verbose(f"    Mapi  file: {self.mapi_json}")
+        Log.verbose(f"  Mapi  file: {self.mapi_json}")
 
     def clean(self, erase: bool):
         if erase:
             Log.resume("Cleaning ", end="")
-            Log.verbose("    Cleaning  : html and cases files")
+            Log.verbose("  Cleaning  : html and cases files")
             os.remove(self.cases)
             os.remove(self.target_html)
 
@@ -136,4 +137,4 @@ class Actions:
     def update_markdown(self):
         if Mdpp.update_file(self.source_readme):
             Log.resume("Mdpp ", end="")
-            Log.verbose(f"    Mdpp updading")
+            Log.verbose(f"  Mdpp updading")