+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- 0.1.4
+  - adicionado modo recursivo no filter
+  - adicionando opcoes de clean, quiet no filter
 - 0.1.3
   - removendo link para css que gerava arquivos html diferentes
 - 0.1.1